cmb::transform_reduce source code and example program complete.

+++ numeric_inner_product.cxx
@@ -8,11 +8,11 @@
 
 int main()
 {
-  std::vector<int> a{ 0, 1, 2, 3, 4 };
-  std::vector<int> b{ 5, 4, 2, 3, 1 };
+  std::vector<double> a{ 0, 1, 2, 3, 4 };
+  std::vector<double> b{ 5, 4, 2, 3, 1 };
 
   auto r1 = cmb::inner_product(a.begin(), a.end(), b.begin(), 0);
-  
+
   auto r2 = cmb::inner_product(a.begin(), a.end(), b.begin(), 0,
                                cmb::plus<>{ }, cmb::equal_to<>{ });
